requirements.txt
camelot 
ghostscript 
fitz
PDFNetPython3
docx2pdf 
io
Pillow
regex